fix(graph): details returns a symbol's whole identity, a slice of its fan-out
Closes #737. `details` is a targeted lookup — the caller named the handle to
learn what it is — so its identity is the answer and is not sampled: a class's
members and a union's values are the symbol itself, and cutting them sends the
model to the file for the rest, the read this index exists to remove. So
`members` and `literals` are uncapped.

Fan-out is the opposite. What names or uses a symbol grows with the symbol's
popularity, not with the symbol: measured on typeorm, an uncapped `DataSource`
came back 390 KB — ~100k tokens — of which 380 KB was its 1060 inbound
type-references, a trace/impact answer nobody asks `details` for. So `calls`,
`types`, and the neighbor lists stay a small orientation slice, with `trace`
for the whole fan-out. `DataSource` is 8 KB now with its 41 members intact;
`EntityMetadata`'s 120 members are 18 KB where reading its 1134-line source is
far more.

This drops the `literalsTruncated` marker #732 shipped with `literals`: the
value set is not a slice to mark, it is returned whole. A `signature` still
stops at its own declaration span rather than reading a neighbor's lines, an
independent correctness fix. The tour is unaffected — it calls `runDetails`
with its own explicit small limits, so its payload is byte-identical, verified
offline against master on typeorm.
